refactor: Update the return type `int` instead of `long` lineNumber
- In real world, 99.99% Use-case file line number does not exceed Integer.MAX_VALUE
- Suppose each line size 1KiB, For exceeding Integer.MAX_VALUE, the file size should be al most 2TB
- Almost case file size is not to exceed 2TB.

+++ source-connector/src/main/java/sourceconnector/service/reader/LineReader.java
@@ -1,6 +1,6 @@
 package sourceconnector.service.reader;
 
 public interface LineReader<T> extends Reader<T> {
-  long getLineNumber();
-  void setLineNumber(long lineNumber);
+  int getLineNumber();
+  void setLineNumber(int lineNumber);
 }
